Extract detail workflow
Move detail entry planning, refresh coordination, runtime authorization, and pending verification state into an embedded workflow. Assign every detail read a monotonic claim so only the newest-started result can update inventory.
