+++ rust/src/manager/cloud_backup_manager/actors/supervisor/enable/recovery.rs
@@ -202,7 +202,7 @@ impl CloudBackupSupervisor {
             return Produces::ok(());
         }
 
-        self.runtime_passkey_authorization = Some(RuntimePasskeyAuthorization {
+        self.detail_workflow.set_authorization(RuntimePasskeyAuthorization {
             namespace_id: namespace_id.clone(),
             credential_id,
             prf_salt,
